backend/requirements.txt
django>=5.2,<6
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
openai
opencv-python
easyocr
pdfplumber
numpy
